[edit] MatchResultをcontainer化、container→containersに変更
