+++ data/readJson.js
@@ -1,4 +1,3 @@
-// jsonファイルをここに追加していく
 export const jsonData = [
   require('./json/20151210_tournament_data_yosen.json'),
   require('./json/20151210_tournament_data_final.json')
